Split dsl object into package
