+++ core/src/main/scala/com/rallyhealth/vapors/core/dsl/package.scala
@@ -1,0 +1,3 @@
+package com.rallyhealth.vapors.core
+
+package object dsl extends FactFilterDsl
